Introduce FancyZonesDataTypes
