+++ src/modules/fancyzones/lib/WindowMoveHandler.cpp
@@ -12,6 +12,7 @@
 #include "VirtualDesktopUtils.h"
 #include "lib/SecondaryMouseButtonsHook.h"
 #include "lib/GenericKeyHook.h"
+#include "lib/FancyZonesData.h"
 
 extern "C" IMAGE_DOS_HEADER __ImageBase;
